Rfid class and MS fw refactoring
